requirements.txt
Flask==2.1.2
pandas==1.3.5
flask-cors==3.0.10
llama-index-llms-ibm==0.1.0

